crypto_pipeline/assets/gold_assets.py
from dagster import asset


@asset(
    name="gold_portfolio_metrics",
    description="[PLACEHOLDER] Portfolio performance analytics",
    group_name="gold_layer",
    compute_kind="spark"
)
def gold_portfolio_metrics():
    """
    Future implementation:
    - Calculate portfolio performance metrics
    - Compute Sharpe ratio, volatility, drawdown
    - Generate risk assessment reports
    """
    return {"status": "placeholder", "layer": "gold", "type": "portfolio"}


@asset(
    name="gold_bigquery_sync",
    description="[PLACEHOLDER] Sync gold layer data to BigQuery",
    group_name="gold_layer",
    compute_kind="bigquery"
)
def gold_bigquery_sync():
    """
    Future implementation:
    - Read from gold Hudi tables
    - Transform for BigQuery schema
    - Sync to BigQuery tables for analytics
    """
    return {"status": "placeholder", "layer": "gold", "type": "bigquery_sync"}
